backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
kiteconnect
psycopg2-binary

